Authorized middleware

+++ routes/api.php
@@ -5,11 +5,16 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AdminLogin;
 
+use App\Http\Middleware\Authorized;
+
 Route::resources([
   'login' => AdminLogin::class,
   'report' => ReportController::class,
-  'reports' => ReportController::class,
 ]);
 
+Route::resource(
+  "/reports",
+  ReportController::class
+)->middleware(Authorized::class);
+Route::delete('/reports/{id}', [ReportController::class, 'destroy'])->middleware([Authorized::class]);
 
-Route::delete('/reports/{id}', [ReportController::class, 'destroy']);
